seperate add player comp form player table
